backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
websockets==14.1
google-generativeai>=0.8.0
beautifulsoup4>=4.12.0
torch>=2.0.0
scikit-learn>=1.3.0
finance-datareader>=0.9.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
